+++ headers/checkFields.h
@@ -8,6 +8,7 @@ CHECK_FOR(ObjectName);
 CHECK_FOR(dead);
 CHECK_FOR(interactWith)
 CHECK_FOR(key);
+CHECK_FOR(DontSerialize);
 
 //macro that makes my getter functions. life is too short brah
 #define make_getter(field,type) \