+++ headers/item.h
@@ -87,6 +87,9 @@ struct PickupComponent
     void collideWith(PhysicsBody& owner, PhysicsBody& other);
     bool isThrown(PhysicsBody& owner); //true if we have been thrown at least once
     void setPickupable(bool val);
+protected:
+    //called when picked up
+    virtual void onPickup(){};
 private:
     bool pickupable = true; //set this to false if you want an object to no longer be pickup-able
 };
@@ -104,6 +107,16 @@ struct Barrel : public Object<RectCollider,TextureRenderer,Barrel,PickupComponen
     }
     void update(Terrain& terrain);
     //void onCollide(PhysicsBody& other);
+};
+
+
+template<>
+struct Factory<Barrel>
+{
+    static constexpr std::string_view ObjectName = "barrel";
+    static constexpr EMPTY_TYPE DontSerialize = {};
+    using Base = FactoryBase<Barrel,
+                            access<Barrel,&Barrel::orient,&Orient::pos>>;
 };
 
 template<typename PhysicsBodyType>
@@ -141,24 +154,114 @@ private:
     bool activated = false; //don't activate until first interaction
 };
 
-struct BarrelReceiver : public Object<RectCollider,TextureRenderer,BarrelReceiver>
+
+//generalist class that does something upon receiving another object
+template<typename T>
+requires IsObject<T>
+struct ObjReceiver : public Object<RectCollider,TextureRenderer,ObjReceiver<T>>
 {
-    std::unique_ptr<OnTrigger> onTrigger;
+    T renderSlave; //keep a copy of the thing we are receiving around, just to render it
     SignalName signal;
 
-    BarrelReceiver()
+    ObjReceiver()
     {
         this->keyVal = 1;
         this->followGravity = true;
-        this->collider.width = 100;
-        this->collider.height = 40;
+        this->collider.width = 150;
+        this->collider.height = 200;
         this->renderer.sprite = Globals::Game.Sprites.getSprite("barrel_receiver.png");
     }
-    void onCollide(PhysicsBody& other);
-    void render();
+    void onCollide(PhysicsBody& other)
+    {
+        if (other.getName() == Factory<T>::ObjectName && !activated)
+        {
+            other.setDead(true);
+            activated = true;
+            Globals::Game.terrain.emitSignal(signal,this);
+            onReceive();
+        }
+    }
+    virtual void onReceive(){};
+    void render()
+    {
+        float rotation = this->getOrient().rotation - M_PI/2;
+        renderSlave.setPos(this->getPos() + Vector2(cos(rotation),sin(rotation))*.25*this->collider.height);
+        renderSlave.setLayer(this->orient.layer);
+        Object<RectCollider,TextureRenderer,ObjReceiver<T>>::render();
+        renderSlave.render();
+    }
+protected:
+    make_setter(activated,bool);
 private:
+    //true if we want this to no longer receive
     bool activated = false;
 };
+
+//wrapper class that allows us to use strings in template parameters
+template<size_t N>
+struct CompileString
+{
+    char data[N];
+    constexpr CompileString(const char(& str)[N])
+    {
+        for (size_t i = 0; i < N; i ++)
+        {
+            data[i] = str[i];
+        }
+    }
+};
+
+/**
+ * @brief generalist factory class for any specialization or descendant of ObjReceiver
+ * 
+ * @tparam T full type of the descendant or specialization
+ * @tparam ObjName compile-time string for the object name
+ */
+template<typename T,CompileString ObjName>
+struct ReceiverFactory
+{
+    static constexpr std::string_view ObjectName = ObjName.data;
+    using Base = FactoryBase<T,
+                            access<T,&T::orient,&Orient::pos>,
+                            access<T,&T::keyVal>,
+                            access<T,&T::signal>>;
+};
+
+using BarrelReceiver = ObjReceiver<Barrel>;
+template<>
+struct Factory<BarrelReceiver> : ReceiverFactory<BarrelReceiver,"barrel_receiver">{};
+
+struct BigGear : public Object<CircleCollider,TextureRenderer,BigGear,PickupComponent>
+{
+    BigGear()
+    {
+        this->followGravity = false; //follows physics once picked up
+        this->collider.radius = 50;
+        this->renderer.sprite = Globals::Game.Sprites.getSprite("gear.png");
+    }
+
+    void onPickup();
+    void update(Terrain& t);
+
+};
+
+template<>
+struct Factory<BigGear>
+{
+    static constexpr char ObjectName[] = "big_gear";
+    using Base = FactoryBase<BigGear,
+                        access<BigGear,&BigGear::orient,&Orient::pos>>;
+};
+
+struct BigGearReceiver : public ObjReceiver<BigGear>
+{
+    using SerializerType = Factory<BigGearReceiver>;
+
+    BigGearReceiver() : ObjReceiver<BigGear>(){};
+    void onReceive();
+};
+
+template<> struct Factory<BigGearReceiver> : ReceiverFactory<BigGearReceiver,"big_gear_receiver">{};
 
 struct TerrainPod :  public Object<CircleCollider,TextureRenderer,TerrainPod,PickupComponent>
 {
@@ -199,24 +302,7 @@ struct Factory<Booster>
                             access<Booster,&Booster::orient,&Orient::pos>>;
 };
 
-/*template<>
-struct Factory<Barrel>
-{
-    static constexpr std::string ObjectName = "barrel";
-    using Base = FactoryBase<Barrel,
-                            access<Barrel,&Barrel::orient,&Orient::pos>>;
-};*/
 
-template<>
-struct Factory<BarrelReceiver>
-{
-    static constexpr std::string ObjectName = "barrel_receiver";
-    using Base = FactoryBase<BarrelReceiver,
-                            access<BarrelReceiver,&BarrelReceiver::orient,&Orient::pos>,
-                            access<BarrelReceiver,&BarrelReceiver::keyVal>,
-                            access<BarrelReceiver,&BarrelReceiver::onTrigger>,
-                            access<BarrelReceiver,&BarrelReceiver::signal>>;
-};
 
 typedef GenericSpawner<Barrel> BarrelSpawner;