+++ src/blocks.cpp
@@ -72,7 +72,6 @@ void Terrain::addBlock(const Vector2& pos, const Block& block, bool draw)
     {
         terrain.resize((index + 1)*TerrainMap::PALETTE_SIZE);
     }
-    terrain.setVal(index,AIR);
     //std::cout << pos << "  " << index << " "<< terrain.size()<< "\n";
 
     Color color;
@@ -134,15 +133,9 @@ void Terrain::remove(const Vector2& pos, int radius)
         rlSetBlendMode(BLEND_CUSTOM);
     //double time = GetTime();
     forEachPos([this](const Vector2& pos){
-            //for (int i = 0; i < 9; i ++)
-            {
-                //Vector2 neighbor = pos;//{pos.x + Block::BLOCK_DIMEN*(i%3 - 1),pos.y + Block::BLOCK_DIMEN*(i/3 - 1)};
-                //terrain.setVal(pointToIndex(neighbor),BlockType::AIR);
-                //DrawRectangle(neighbor.x*PIXEL_RATIO,(blocksTexture.texture.height - neighbor.y)*PIXEL_RATIO -  PIXEL_SIZE,PIXEL_SIZE,PIXEL_SIZE,Fade(BLACK, 0.0f));
 
-                addBlock(pos,{BlockType::AIR},true);
+                addBlock(pos,{WHITE,BlockType::AIR},true);
 
-            }
                },pos,radius+Block::BLOCK_DIMEN*2*(PIXEL_RATIO));
         rlSetBlendMode(BLEND_ALPHA);
     endDrawBlocks();
@@ -236,6 +229,7 @@ void Terrain::generatePlanet(const Vector2& center, int radius, const Color& col
     forEachPos([this,color](const Vector2& pos){
                     addBlock(pos,{color});
                },center,radius);
+
     endDrawBlocks();
 
 }
@@ -313,7 +307,7 @@ bool Terrain::checkBlocks(const Vector2& pos, bool checkPlanets, std::function<b
     {
         for (auto it = planets.begin(); it != planets.end();)
         {
-            if (it->ptr.lock() && check(it->type) && CheckCollisionPointShape(pos,it->ptr.lock()->getShape()))
+            if (it->ptr.lock() && it->ptr.lock()->isTangible() && check(it->type)  && CheckCollisionPointShape(pos,it->ptr.lock()->getShape()))
             {
                 return true;
             }
@@ -440,7 +434,7 @@ Vector2 Terrain::lineBlockIntersect(const Vector2& a, const Vector2& b, CheckFun
     Vector2 answer = past ? b : current;
     for (EntityPlanet& terr : planets) //calculate the answer completely separately, by only accounting for planets
         {
-            if (terr.ptr.lock() && check(terr.type))
+            if (terr.ptr.lock() && terr.ptr.lock()->isTangible() && check(terr.type))
             {
                 Vector2 intersect = lineShapeIntersect(terr.ptr.lock()->getShape(),a,b);
                 if (Vector2DistanceSqr(intersect,a) < Vector2DistanceSqr(a,b))