+++ src/factory.cpp
@@ -36,6 +36,11 @@ void ClassDeserializer::init()
 
     registerName<LifePod>();
     registerName<Battery>("battery");
+
+    registerName<Disintegrate>();
+
+    registerName<BigGear>();
+    registerName<BigGearReceiver>();
 }
 
 std::shared_ptr<PhysicsBody> ClassDeserializer::construct(std::string_view cereal)