+++ CHANGELOG.md
@@ -5,6 +5,28 @@
 * TODO: Make ESCAPE open a pause menu
 * TODO: Make ALT-TAB work
 * TODO: reverse gravity is cool, planets need some kind of particle effect pushing away to very obviously show where the field is strong, its not intuitive like normal gravity
+* TODO: Make SphereTerrain actaully render something
+* TODO: Fix bouncing on terrain bug
+  * Possibly related: Fix being able to walk off terrain
+* Finally fix portals rendering improperly
+* Add Andrew's music
+
+## 6/22/2026 Messing Around, Finalizing Levels
+* Added `Disintegrate` a `CircleTerrain` that disappears for a time when stepped on
+* Added `BigGear`, a larger, carryable gear that needs to be carried to a receiver
+* Added the `ObjReceiver`, an object that does something when receiving an object, basically a generalized `BarrelReceiver`
+  * Added `BigGearReceiver`
+* `LaserBeamer` now saves its starting location, preventing them from wandering every time we save
+* Made some slightly better docs for `accessSetter`
+* `GravityStream` and `GravitySwitch` have been reworked (again)
+  * `GravitySwitch` is no longer used
+  * `GravityStream` is now a constant stream of force that moves a player in a specific direction. 
+* Added the `HOLE` block type, which has 20% more gravity. Currently unused
+* Added `onDeserialize`, a function to be called specifically when an object is added in the editor 
+* Removed some of the barrel tossing puzzles. They honestly feel lame af.
+* Added a convenience portal to layer 4
+* Finished levels for demo
+
 
 ## 5/26/2026 World Map UI Updates
 * World Map now has a gradient ellipse for each level